Fixed problem with ros-build failing due to the package.in.xml not being converted to package.xml before bloom attempts to parse it.
Unlike colcon, bloom parses package.xml before invoking CMake, so it doesn't get filled in first. The current solution is to just fill it in ourselves using sed, but a better solution probably exists. In the future, it may be worth investigating abandoning bloom and packaging using colcon/CPack ourselves.